Fix theme loading and add function to remove bg
Removing the BG color can be nice if you've got special colors in a terminal,
but it can also obscure the text, so it's a function that's not enabled by
default.  I could capture the current value of the BG color and make this a
toggle, but I've spent enough time on config for today.